Set files to be patched to 0644 if they are 0444, set back after.
Issue 31:
In an environment that uses the Lustre file system, it's been seen
that any files that are set to 0444 (read only) will cause fatal
filesystem errors when the patch is attempted. This makes using
perlbrew to install perl on these systems impossible This change
detects read-only files about to be patched, sets them to 0644
for the duration of the patching, then sets them back to 0444 so
that they exist as expected during the actual build process.