fix darwin DynaLoader patch for older perls
The patch that was taken from blead uses a <<$^OS-eq-darwin>>
preprocessing token, which is handled by the DynaLoader_pm.PL script.
Older versions of DynaLoader_pm.PL don't include this preprocessing, but
the patch still applies.

In our case, the preprocessing is a check for darwin. But the patch is
already only applied on darwin, so it is redundant. We can just remove
the preprocessing token, and it should work across all versions.